feature: Refactor Header component into NavBar and MobileNav subcomponents
