+++ src/components/Header/Header.module.scss
@@ -4,89 +4,3 @@
     align-items: center;
     padding-inline: 20px;
 }
-
-.navbar {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 100%;
-
-    svg {
-        cursor: pointer;
-    }
-
-    .menuIcon {
-        width: 24px;
-        height: 24px;
-    }
-
-
-    .logoIcon {
-        width: 32px;
-        height: 32px;
-    }
-
-    .iconGroup {
-        display: flex;
-        flex-direction: row;
-        gap: 20px;
-
-        svg {
-            width: 20px;
-            height: 20px;
-        }
-    }
-
-}
-
-.mobileNav {
-    display: flex;
-    flex-direction: column;
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    opacity: 0;
-    visibility: hidden;
-    background-color: var(--background);
-    z-index: 10;
-
-    transition: opacity 0.3s ease-in-out, visibility 0.3s ease-in-out;
-
-    &.open {
-        opacity: 1;
-        visibility: visible;
-    }
-
-    .mobileNavHeader {
-        height: var(--header-height);
-        display: flex;
-        align-items: center;
-        padding-inline: 20px;
-        margin-bottom: 21px;
-
-        .closeIcon {
-            width: 24px;
-            height: 24px;
-            padding: 5px;
-            cursor: pointer;
-        }
-    }
-
-    .linkList {
-        display: flex;
-        flex-direction: column;
-        gap: 20px;
-        padding-inline: 20px;
-        padding-block: 0px 30px;
-        border-bottom: 1px solid var(--clr-gray-90);
-
-    }
-
-    .linkItem {
-        font-weight: 200;
-        font-size: 18px;
-    }
-
-
-}
