+++ src/components/Header/MobileNav.module.scss
@@ -1,0 +1,51 @@
+.mobileNav {
+    display: flex;
+    flex-direction: column;
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    opacity: 0;
+    visibility: hidden;
+    background-color: var(--background);
+    z-index: 10;
+
+    transition: opacity 0.3s ease-in-out, visibility 0.3s ease-in-out;
+
+    &.open {
+        opacity: 1;
+        visibility: visible;
+    }
+
+    .mobileNavHeader {
+        height: var(--header-height);
+        display: flex;
+        align-items: center;
+        padding-inline: 20px;
+        margin-bottom: 21px;
+
+        .closeIcon {
+            width: 24px;
+            height: 24px;
+            padding: 5px;
+            cursor: pointer;
+        }
+    }
+
+    .linkList {
+        display: flex;
+        flex-direction: column;
+        gap: 20px;
+        padding-inline: 20px;
+        padding-block: 0px 30px;
+        border-bottom: 1px solid var(--clr-gray-90);
+
+    }
+
+    .linkItem {
+        font-weight: 200;
+        font-size: 18px;
+    }
+
+
+}
