+++ src/components/Header/NavBar.module.scss
@@ -1,0 +1,32 @@
+.navbar {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    width: 100%;
+
+    svg {
+        cursor: pointer;
+    }
+
+    .menuIcon {
+        width: 24px;
+        height: 24px;
+    }
+
+    .logoIcon {
+        width: 32px;
+        height: 32px;
+    }
+
+    .iconGroup {
+        display: flex;
+        flex-direction: row;
+        gap: 20px;
+
+        svg {
+            width: 20px;
+            height: 20px;
+        }
+    }
+
+}
